fix updating data in Home vc

+++ EwanoCash/TransferPage/Model.swift
@@ -9,7 +9,7 @@ import Foundation
 
 struct TransactionData: Codable, Equatable {
     var title = ""
-    var amount: Double = 0
+    var amount: Int = 0
     var date: Date
     var isIncome = true
 }
